remove like id (not needed): is now a composite entity + migration
