+++ Service.Contracts/IServiceManager.cs
@@ -3,7 +3,6 @@
     public interface IServiceManager
     {
         IMovieService MovieService { get; }
-        ILikeService LikeService { get; }
         IUserService UserService { get; }
         ICinemaService CinemaService { get; }
     }
