+++ Service.Contracts/IUserService.cs
@@ -7,4 +7,7 @@ public interface IUserService
     UserDto CreateUser(UserForCreationDto user);
     UserDto GetUserById(Guid id, bool trackChanges);
     UserDto GetUserByEmail(string email, bool trackChanges);
+    bool IsMovieLikedByUser(Guid userId, Guid movieId, bool trackChanges);
+    LikeDto CreateLike(LikeDto likeDto);
+
 }
