+++ Shared/DataTransferObjects/postDTO.cs
@@ -1,12 +1,3 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Shared.DataTransferObjects;
+﻿namespace Shared.DataTransferObjects;
 
 public record UserForCreationDto(string Email, string FullName);
-
-public record LikeForCreationDto(string MovieUuid, string UserUuid);
